+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/template/UITemplateInitializer.java
@@ -18,13 +18,13 @@
 
 package org.apache.skywalking.oap.server.core.management.ui.template;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.File;
 import java.io.IOException;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.skywalking.oap.server.core.CoreModule;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
@@ -35,20 +35,19 @@ import org.apache.skywalking.oap.server.library.util.ResourceUtils;
 
 /**
  * UITemplateInitializer load the template from the config file in json format. It depends on the UI implementation only.
- * Each config file should has only one setting json object.
- * The configs should has different names in the same Layer and entity.
+ * Each config file should be only one dashboard setting json object.
+ * The dashboard names should be different in the same Layer and entity.
  */
 @Slf4j
 public class UITemplateInitializer {
-    public static Layer[] supportedLayer = new Layer[] {
+    public static Layer[] SUPPORTED_LAYER = new Layer[] {
         Layer.MESH,
         Layer.GENERAL,
         Layer.K8S,
         Layer.BROWSER,
         Layer.SO11Y_OAP
     };
-    private Map yamlData;
-    private UITemplateManagementService uiTemplateManagementService;
+    private final UITemplateManagementService uiTemplateManagementService;
 
     public UITemplateInitializer(ModuleManager manager) {
         this.uiTemplateManagementService = manager.find(CoreModule.NAME)
@@ -56,13 +55,8 @@ public class UITemplateInitializer {
                                                   .getService(UITemplateManagementService.class);
     }
 
-    public UITemplateInitializer() {
-
-    }
-
     public void initAll() throws IOException {
-        for (Layer layer : UITemplateInitializer.supportedLayer) {
-            System.out.println();
+        for (Layer layer : UITemplateInitializer.SUPPORTED_LAYER) {
             File[] templateFiles = ResourceUtils.getPathFiles("ui-initialized-templates/" + layer.name().toLowerCase(
                 Locale.ROOT));
             for (File file : templateFiles) {
@@ -73,27 +67,23 @@ public class UITemplateInitializer {
 
     public void initTemplate(File template) throws IOException {
         ObjectMapper mapper = new ObjectMapper();
+        mapper.enable(JsonParser.Feature.ALLOW_COMMENTS);
         JsonNode jsonNode = mapper.readTree(template);
         if (jsonNode.size() > 1) {
             throw new IllegalArgumentException(
-                "UI template " + template.getName() + "should has only one setting json object.");
+                "File:  " + template.getName() + " should be only one dashboard setting json object.");
         }
-        JsonNode configNode = jsonNode.get(0);
+        JsonNode configNode = jsonNode.get(0).get("configuration");
         String inId = configNode.get("id").textValue();
         String inName = configNode.get("name").textValue();
 
         verifyNameConflict(template, inId, inName);
-        //Todo: implement others validation.
 
         DashboardSetting setting = new DashboardSetting();
-        setting.setUpdateTime(System.currentTimeMillis());
-        setting.setId(configNode.get("id").textValue());
-        setting.setConfiguration(configNode.get("configuration").toString());
+        setting.setId(inId);
+        setting.setConfiguration(configNode.toString());
 
-        uiTemplateManagementService.getTemplate(inId);
-
-        uiTemplateManagementService.addTemplate(setting);
-
+        uiTemplateManagementService.addOrUpdate(setting);
     }
 
     private void verifyNameConflict(File template, String inId, String inName) throws IOException {
@@ -104,7 +94,7 @@ public class UITemplateInitializer {
             String id = jsonNode.get("id").textValue();
             if (jsonNode.get("name").textValue().equals(inName) && !id.equals(inId)) {
                 throw new IllegalArgumentException(
-                    "UI template " + template.getName() + "name: " + inName + "conflict with exist configuration id: " + id);
+                    "File:  " + template.getName() + " name: " + inName + " conflict with exist configuration id: " + id);
             }
         }
     }