+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/GraphQLQueryConfig.java
@@ -30,4 +30,5 @@ import org.apache.skywalking.oap.server.library.module.ModuleConfig;
 public class GraphQLQueryConfig extends ModuleConfig {
     private boolean enableLogTestTool;
     private int maxQueryComplexity = 100;
+    private boolean enableUpdateUITemplate = true;
 }